Fix overflow during validation

+++ runtime/src/valid.rs
@@ -498,9 +498,9 @@ fn check_table(table: &ast::Table) -> Option<()> {
 
 fn check_memory(mem: &ast::Memory) -> Option<()> {
     // Can't allocate more than 4GB since its a 32-bits machine
-    let max = (1 << 32) / 65536;
-    if mem.type_.limits.min as usize > max
-        || (mem.type_.limits.max.is_some() && mem.type_.limits.max.unwrap() as usize > max)
+    let max = (1u64 << 32) / 65536;
+    if mem.type_.limits.min as u64 > max
+        || (mem.type_.limits.max.is_some() && mem.type_.limits.max.unwrap() as u64 > max)
     {
         return None;
     }
